+++ app/src/main/java/github/karchx/wiki/listeners/ArticleItemClickListener.kt
@@ -1,3 +1,8 @@
+/*
+ * Copyright 2021 Andrey Karchevsky <karch.andrus@gmail.com>
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
 package github.karchx.wiki.listeners
 
 import android.content.Context
