+++ app/src/main/res/drawable/recycle_view_divider.xml
@@ -1,4 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
+<!--
+  ~ Copyright 2021 Alex Syrnikov <pioneer19@post.cz>
+  ~ SPDX-License-Identifier: Apache-2.0
+  -->
 <shape xmlns:android="http://schemas.android.com/apk/res/android"
     android:shape="rectangle">
     <size
