+++ app/src/main/res/layout/search_fragment.xml
@@ -1,4 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
+<!--
+  ~ Copyright 2021 Andrey Karchevsky <karch.andrus@gmail.com>
+  ~ SPDX-License-Identifier: Apache-2.0
+ -->
 <androidx.constraintlayout.widget.ConstraintLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
@@ -19,14 +23,14 @@
             android:id="@+id/editTextUserRequest"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:hint="Search in Wikipedia" />
+            android:hint="@string/search_in_wikipedia" />
     </com.google.android.material.textfield.TextInputLayout>
 
     <Button
         android:id="@+id/searchButton"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
-        android:text="Button"
+        android:text="@string/find_button"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/textInputLayoutUserRequest" />
