+++ app/src/main/res/navigation/nav_graph.xml
@@ -1,4 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
+<!--
+  ~ Copyright 2021 Alex Syrnikov <pioneer19@post.cz>
+  ~ SPDX-License-Identifier: Apache-2.0
+  -->
 <navigation xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
