+++ app/src/main/res/values/strings.xml
@@ -6,4 +6,6 @@
     <string name="empty_field_error">This field cannot me empty</string>
     <string name="incorrect_request_error">Incorrect request. Nothing found</string>
     <string name="internet_connection_error">No internet, check your connection</string>
+    <string name="search_in_wikipedia">Search in Wikipedia</string>
+    <string name="find_button">Find</string>
 </resources>
